Cleanup: move computeRetryPolicy to WebClientEventSender
